fix(inode): name libc::stat's field widths per target, and refuse an unknown one
`libc::stat`'s fields are not one shape. Within Linux, `nlink_t` is 64-bit on x86_64
and 32-bit on aarch64 and `blksize_t` follows; darwin narrows `mode_t` and `nlink_t` to
16-bit and makes `dev_t` a *signed* 32-bit. So `to_stat` now spells the five fields
whose width varies out per target, with the widths for one target kept together, and
the fields that are the same everywhere -- st_ino, st_uid, st_gid, st_size, st_blocks
and the times -- assigned once above them.

A target that is not listed no longer builds. The previous form named x86_64 and
aarch64 and stopped, so on any other architecture neither arm compiled and `st_nlink`
and `st_blksize` kept the zero from the zeroed initialiser, reporting a file with no
links and no block size. Nothing said so, and a 10-line probe built for riscv64
confirms it: the compiler reports the field as never read. This is a behaviour change
for such a target -- it fails to build where it used to compile and answer wrongly.

`update_stat` needs no arms at all, and the asymmetry is worth naming: that direction
converts *into* the inode's own fixed-width fields, so every host's 64-, 32- or 16-bit
value widens with the same cast. Only handing out the host's own struct has a moving
target.

The mode bits are cast once per expression rather than at each constant, with `S_IFMT`
given a name in the width `i_mode` stores. Their *values* need nothing: S_IFMT, the
S_IF* types and the permission bits are identical on both platforms, so a mode that
goes to storage means the same thing read back from either -- only the type is
narrower. Which is the point: the on-disk inode is fixed-width, so the host's `stat`
widths never reach it.

+++ src/file/file.rs
@@ -889,7 +889,8 @@ impl<'a, T, L, C> HyperFile<'a, T, L, C>
         let raw = inode.to_raw(raw_inode.i_bmap);
         let od_state = inode.get_ondisk_state();
         let _ = staging.flush_inode(raw.as_u8_slice(), od_state, FlushInodeFlag::Update).await?;
-        Ok(inode.to_stat(stat.st_dev, stat.st_rdev))
+        // `to_stat` takes the inode's own width; `dev_t` is a signed 32-bit on darwin.
+        Ok(inode.to_stat(stat.st_dev as u64, stat.st_rdev as u64))
     }
 
     pub async fn update_stat(&mut self, stat: &libc::stat) -> Result<libc::stat> {
